Cleanup cpp examples
